Move iterators and iterables to kotlinutil
